Major changes to the Whitelist and Cleaner.
Modifications to handle ids and attributes in the whitelist.
Whitelist returns null instead of the default block when the element is
not found in the whitelist.

The Cleaner normalizes inline elements.
The Cleaner removes non-whitelisted blocks unless its children are all
inline nodes, in which case it replaces the block with the default block.

Removes spans from the whitelist, only allowing RANGE_START and RANGE_END.