Updates the EraseHandler.
The EraseHandler now has a config so that selectors can be passed
through. These selectors are used to determine if the element to
be deleted should be completely deleted or merged. This gives a
greater flexibility and removes the atomic hack as this new solution
is general.

This also solves the issue in FF, IE8/9 where you can't delete an hr.
The hr is not added to the delete selectors and the Erasehandler just
deletes it completely when it sees an hr.