Adds onInclude() to buttons.
Moves the instantiation of the API to above the includes.
Adds includeButtons().
Adds a way for shortcuts to be related to buttons through actions.
Moves inclusion of shortcuts and behaviours into the onInclude().
Modifies toolbar to take the editor instead of an api.
Updates plugin Helpers.createButton() to default the action to the
name of the button and add an onInclude for shortcuts and behaviours.