Fixes #185 - IE8: Clicking on an image to activate the SnapEditor breaks
The cleaner was replacing the image so when keepRange() was trying to
reselect the image for the range, it no longer existed and broke.

Also, the image wasn't technically supposed to be replaced. IE8 was
adding and extra "complete" attribute that was identified as user-defined.
This caused an extra attribute in the element which wasn't in the
whitelist. This triggered the replacement code. This has been fixed as
well.