Fixes #188 - Don't clean on initialize
The cleaner no longer cleans on initialize.

The editor triggers a snapeditor.before_activate event for the Save
plugin to grab the HTML before any plugins get a chance to do anything.

The Save plugin grabs the raw HTML from the element instead of the
cleaned HTML from SnapEditor.

getContents() returns the content without cleaning out zwnbsp. This
feels more inline with the true content that SnapEditor is displaying
so that it always matches the content when viewing without SnapEditor.