Adds #199 - Make the save plugin more customizable
Cleans up the tryDeactivate flow so that the code to override the default is
in the save plugin.
Leaves save() in the save plugin instead of the API object.
Save plugin adds an extra onUnsavedChanges to override the default flow. The
user can now use api.execAction() and pass in "save" or "discard" to access
the save plugin's actions. Originally I thought of adding this to the API, but
this way is more in line with the plugins.