Fixes #201 - iframe access causes security errors
Instead of listening to the inside of iframes, we overlay shims to cover up
the iframes. This prevents any sort of iframe access.

Refactored Form stuff into the Form instead of leaving it in the Base.

Changed transformCoordinatesRelativeToOuter to take in an iframe instead of a
target to prevent security issues.

Removes ContextMenu as it is no longer needed.