Fixes #239 - Firefox shows errors when typing in SnapEditor.Form.
This is not just an error with SnapEditor.Form but was more general. The bug
appears when the cursor is not collapsed at the end of the document and with
nothing else after it. There were no returned client rectangles in this case
because the temporary range created was still collapsed since it was at the
end. This case happens more likely in SnapEditor.Form because it is inside an
iframe and the SnapEditor text is the only content. This could theoretically
happen in SnapEditor.InPlace if there was nothing else after it.

The fix is it use the third scenario where we use a span to calculate the
coordinates.