Fixes #251 - Style newline cannot reference itself
The issue was that when a particular style had never been added yet
and the newline referenced itself, the newline gets parsed and created
first before the style itself. This created 2 styles with the newline
being the first one created. This caused problems when looking up the
whitelist because the newline style would be returned first, which
didn't have a newline.

This felt wrong anyways since newlines should be styles that are
whitelisted instead of newline styles being added to the whitelist.

The fix was to keep newlines as a label or rule instead of parsing
them and adding them to the whitelist. When needed, the newline rule
would be parsed at the time and the matching whitelist object would
be found.